// frontend/packages/blockly/src/types.ts
export interface ThemeBlockStyle {
  colourPrimary: string;
  colourSecondary?: string;
  colourTertiary?: string;
  hat?: string;
}

/**
 * Our custom theme interface.
 */
export interface Theme {
  definition: {
    /**
     * The human-oriented name of the Theme to place in any settings dialog.
     */
    option: string;
    blockLimits: {
      indicator: {
        fill: string;
        text: string;
      };
      overLimit: {
        fill: string;
        text: string;
      };
    };
    trashcan: {
      fill: string;
    };
    blockStyles: {
      default: ThemeBlockStyle;
      setup_blocks: ThemeBlockStyle;
      event_blocks: ThemeBlockStyle;
      loop_blocks: ThemeBlockStyle;
      logic_blocks: ThemeBlockStyle;
      procedure_blocks: ThemeBlockStyle;
      variable_blocks: ThemeBlockStyle;
      math_blocks: ThemeBlockStyle;
      text_blocks: ThemeBlockStyle;
      colour_blocks: ThemeBlockStyle;
      // Used in Sprite Lab, Dance, Poetry
      sprite_blocks: ThemeBlockStyle;
      // Used in Sprite Lab, Dance, Poetry
      world_blocks: ThemeBlockStyle;
      // Used in Sprite Lab, Dance
      behavior_blocks: ThemeBlockStyle;
      // Used in Sprite Lab only
      location_blocks: ThemeBlockStyle;
      // Formerly called dance_blocks, music_blocks
      lab_blocks: ThemeBlockStyle;
    };
  } & Omit<
    InstanceType<typeof Blockly.Theme>,
    | 'getClassName'
    | 'setBlockStyle'
    | 'setCategoryStyle'
    | 'getComponentStyle'
    | 'setComponentStyle'
    | 'setFontStyle'
    | 'setStartHats'
    | 'blockStyles'
  >;
  instance: Blockly.Theme;
}